Moved Model Testing into their own folder and added more error handling to my dashboard method
