+++ src/NUnit_Tests/ModelTesting/GameTests.cs
@@ -1,4 +1,3 @@
-using NUnit_Tests.RepoTesting;
 using SteamProject.Models;
 
 namespace NUnit_Tests.ModelTesting
