+++ src/NUnit_Tests/ModelTesting/UserLibraryVMTests.cs
@@ -1,4 +1,3 @@
-using NUnit_Tests.RepoTesting;
 using SteamProject.Models;
 using SteamProject.ViewModels;
