+++ src/NUnit_Tests/ModelTesting/UserTests.cs
@@ -1,4 +1,3 @@
-using NUnit_Tests.RepoTesting;
 using SteamProject.Models;
 using SteamProject.Models.DTO;
 using System.Text.Json;
